refactor(settings): decompose SettingsService into Flags/Write/PettyCash/DocNumberPreview services
Behavior-preserving. Facade keeps the 15-method public surface + delegates to 4
sub-services; getUiFlags (~800 LOC) + the bulkUpdate $transaction moved VERBATIM.
DocNumberPreview injects SettingsFlagsService for getKey/getDocPrefixMap. Facade
+ controller + external consumers unaffected. tests green, tsc 0, lint 0.
