+++ apps/api/src/modules/settings/__tests__/repair-config-defaults.spec.ts
@@ -14,6 +14,10 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { Test, TestingModule } from '@nestjs/testing';
 import { SettingsService } from '../settings.service';
+import { SettingsFlagsService } from '../services/settings-flags.service';
+import { SettingsWriteService } from '../services/settings-write.service';
+import { PettyCashCustodianService } from '../services/petty-cash-custodian.service';
+import { DocNumberPreviewService } from '../services/doc-number-preview.service';
 import { PrismaService } from '../../../prisma/prisma.service';
 import { AuditService } from '../../audit/audit.service';
 
@@ -61,6 +65,10 @@ describe('Repair ticket SystemConfig defaults', () => {
       const mod: TestingModule = await Test.createTestingModule({
         providers: [
           SettingsService,
+          SettingsFlagsService,
+          SettingsWriteService,
+          PettyCashCustodianService,
+          DocNumberPreviewService,
           { provide: PrismaService, useValue: prisma },
           { provide: AuditService, useValue: audit },
         ],
