+++ apps/api/src/modules/settings/services/settings-flags.service.ts
@@ -1,0 +1,1030 @@
+import { Injectable } from '@nestjs/common';
+import { DocumentType } from '@prisma/client';
+import { PrismaService } from '../../../prisma/prisma.service';
+import { readBoolFlag, readNumberFlag } from '../../../utils/config.util';
+import { pickEnum, clampInt, clampFloat, parseStringArray } from '../settings-flag-parsers.util';
+import { DEFAULT_DOC_PREFIX_MAP, DOC_PREFIX_REGEX } from '../settings.constants';
+
+/**
+ * Read-mostly SystemConfig flag accessors decomposed out of the monolithic
+ * SettingsService (Wave-4). This is the dependency ROOT for the settings
+ * cluster — `getKey` (made public here) is consumed by DocNumberPreviewService.
+ * All method bodies are byte-identical to the original; only `this.prisma`
+ * field resolution + import paths changed.
+ */
+@Injectable()
+export class SettingsFlagsService {
+  constructor(private prisma: PrismaService) {}
+
+  /**
+   * Read a single SystemConfig value by key. Returns null when key missing.
+   * Used by typed accessors below; not exposed through the controller.
+   */
+  async getKey(key: string): Promise<string | null> {
+    const row = await this.prisma.systemConfig.findFirst({
+      where: { key, deletedAt: null },
+      select: { value: true },
+    });
+    return row?.value ?? null;
+  }
+
+  // Delegate to shared util — keeps parsing semantics identical across services.
+  private async readNumber(key: string, fallback: number): Promise<number> {
+    return readNumberFlag(this.prisma, key, fallback);
+  }
+
+  private async readBoolean(key: string, fallback: boolean): Promise<boolean> {
+    return readBoolFlag(this.prisma, key, fallback);
+  }
+
+  /**
+   * D1.3.3.1 — public accessor for the export-enabled flag, used by export
+   * endpoints (PDF receipts, trade-in vouchers, OI receipt, reporting PDF)
+   * to short-circuit with 403 before any heavy generation happens. Defaults
+   * to true so existing exports keep working when the SystemConfig row is
+   * absent (first-boot / fresh-DB).
+   */
+  async isExportEnabled(): Promise<boolean> {
+    return this.readBoolean('export_enabled', true);
+  }
+
+  /**
+   * D1.1.3.2 — DB-driven WHT-rate dropdown. JSON-encoded array of
+   * `{rate, label}` objects stored in SystemConfig key `wht_rates`.
+   * Default = the 5 canonical rates (1/3/5/10/15 %).
+   *
+   * D1.1.3.5 — each entry MAY also carry an optional `effectiveDate`
+   * (ISO-8601 string). `null`/missing means "always active". Frontend
+   * filters out future-dated entries client-side when rendering the
+   * dropdown. The server keeps stored history intact.
+   *
+   * Validation rules per entry:
+   *   - `rate` must be a finite number in [0, 30]
+   *   - `label` must be a non-empty string
+   *   - `effectiveDate`, if present, must be a string that parses to a
+   *     valid Date (Date.parse not NaN)
+   * If ANY entry fails validation, fall back to defaults wholesale —
+   * partial/malformed data leaks confusing UI options.
+   */
+  async getWhtRates(): Promise<
+    { rate: number; label: string; effectiveDate?: string | null }[]
+  > {
+    const raw = await this.getKey('wht_rates');
+    const defaults: { rate: number; label: string; effectiveDate?: string | null }[] = [
+      { rate: 1, label: '1% — ดอกเบี้ย' },
+      { rate: 3, label: '3% — ค่าบริการ' },
+      { rate: 5, label: '5% — ค่าเช่า' },
+      { rate: 10, label: '10% — ค่าวิชาชีพ' },
+      { rate: 15, label: '15% — ต่างประเทศ' },
+    ];
+    if (!raw) return defaults;
+    try {
+      const parsed = JSON.parse(raw);
+      if (
+        !Array.isArray(parsed) ||
+        parsed.length === 0 ||
+        !parsed.every(
+          (r) =>
+            r &&
+            typeof r === 'object' &&
+            typeof r.rate === 'number' &&
+            Number.isFinite(r.rate) &&
+            r.rate >= 0 &&
+            r.rate <= 30 &&
+            typeof r.label === 'string' &&
+            r.label.trim().length > 0 &&
+            // effectiveDate is optional — null/undefined or a parsable string.
+            (r.effectiveDate == null ||
+              (typeof r.effectiveDate === 'string' &&
+                !Number.isNaN(Date.parse(r.effectiveDate)))),
+        )
+      ) {
+        return defaults;
+      }
+      // Normalize: drop unknown keys, coerce effectiveDate to string|null.
+      return parsed.map((r) => ({
+        rate: r.rate,
+        label: r.label,
+        ...(r.effectiveDate ? { effectiveDate: r.effectiveDate as string } : {}),
+      }));
+    } catch {
+      return defaults;
+    }
+  }
+
+  /**
+   * D1.2.7.2 + InternalControlActionBar — DB-driven reverse-reason dropdown.
+   *
+   * Source of truth (in priority order):
+   *   1. `reverse_reasons` table (admin-managed via `/settings#internal-control`).
+   *      Each row maps to `{ code: row.id, label: row.label }`. Inactive
+   *      rows are excluded; sort by `(sortOrder, createdAt)`.
+   *   2. `SystemConfig` key `reverse_reasons` (legacy JSON blob). Still
+   *      honored when the table is empty so partially-migrated environments
+   *      keep working.
+   *   3. Built-in defaults (6 canonical reasons).
+   *
+   * Caller is responsible for treating an empty/invalid stored list as
+   * "use default" — never as "empty list".
+   */
+  async getReverseReasons(): Promise<{ code: string; label: string }[]> {
+    const defaults: { code: string; label: string }[] = [
+      { code: 'data_entry_error', label: 'ป้อนข้อมูลผิด' },
+      { code: 'wrong_vendor', label: 'ผู้ขายผิด' },
+      { code: 'wrong_amount', label: 'จำนวนเงินผิด' },
+      { code: 'duplicate_entry', label: 'ข้อมูลซ้ำ' },
+      { code: 'cancel_transaction', label: 'ยกเลิกรายการ' },
+      { code: 'other', label: 'อื่นๆ (ระบุรายละเอียด)' },
+    ];
+
+    // (1) Try the admin-managed table first.
+    try {
+      const rows = await this.prisma.reverseReason.findMany({
+        where: { isActive: true, deletedAt: null },
+        orderBy: [{ sortOrder: 'asc' }, { createdAt: 'asc' }],
+        select: { id: true, label: true },
+      });
+      if (rows.length > 0) {
+        return rows.map((r) => ({ code: r.id, label: r.label }));
+      }
+    } catch {
+      // table missing (pre-migration) or other DB hiccup → fall through.
+    }
+
+    // (2) Legacy SystemConfig JSON blob.
+    const raw = await this.getKey('reverse_reasons');
+    if (!raw) return defaults;
+    try {
+      const parsed = JSON.parse(raw);
+      if (
+        Array.isArray(parsed) &&
+        parsed.length > 0 &&
+        parsed.every(
+          (r) =>
+            r && typeof r === 'object' && typeof r.code === 'string' && typeof r.label === 'string',
+        )
+      ) {
+        return parsed;
+      }
+      return defaults;
+    } catch {
+      return defaults;
+    }
+  }
+
+  /**
+   * D1.1.2.1 — DocumentType → prefix mapping. Reads SystemConfig key
+   * `doc_prefix_per_type` (JSON object). Falls back to `DEFAULT_DOC_PREFIX_MAP`
+   * when the key is missing, malformed, or any value fails the
+   * `DOC_PREFIX_REGEX` (2-4 uppercase Latin letters).
+   *
+   * Partial overrides are supported: a stored `{ "EXPENSE": "EXP" }` overrides
+   * the EXPENSE prefix and falls back to defaults for every other type.
+   *
+   * **Safety**: invalid stored values do NOT throw at read time — they're
+   * silently replaced with the default so doc creation never blocks on a bad
+   * SystemConfig row. The validation guard in `bulkUpdate` rejects malformed
+   * values at write time.
+   */
+  async getDocPrefixMap(): Promise<Record<DocumentType, string>> {
+    const raw = await this.getKey('doc_prefix_per_type');
+    if (!raw) return { ...DEFAULT_DOC_PREFIX_MAP };
+    try {
+      const parsed = JSON.parse(raw);
+      if (!parsed || typeof parsed !== 'object' || Array.isArray(parsed)) {
+        return { ...DEFAULT_DOC_PREFIX_MAP };
+      }
+      const result: Record<DocumentType, string> = { ...DEFAULT_DOC_PREFIX_MAP };
+      for (const key of Object.keys(DEFAULT_DOC_PREFIX_MAP) as DocumentType[]) {
+        const candidate = (parsed as Record<string, unknown>)[key];
+        if (typeof candidate === 'string' && DOC_PREFIX_REGEX.test(candidate)) {
+          result[key] = candidate;
+        }
+      }
+      return result;
+    } catch {
+      return { ...DEFAULT_DOC_PREFIX_MAP };
+    }
+  }
+
+  async getUiFlags(): Promise<{
+    /** D1.2.8.2 — show ม.42 tax-exempt warning when a payroll custom-income line is marked non-taxable. Default true. */
+    taxExemptWarningEnabled: boolean;
+    /** D1.2.7.1 — require reason on void/reverse dialog. Default true. */
+    reverseReasonRequired: boolean;
+    /** D1.2.7.2 — configurable reverse-reason dropdown. Always at least the 6 defaults. */
+    reverseReasons: { code: string; label: string }[];
+    /** D1.2.6.4 — allow forward-dated transactions (reverse JE / payment / etc.). Default true. */
+    paymentDateAllowFuture: boolean;
+    /** D1.2.2.7 — render verification QR code on voucher footers. Default true. */
+    voucherShowQrCode: boolean;
+    /**
+     * D1.2.2.6 — UI language preference. 'th' default. The web app applies
+     * the value to `document.documentElement.lang` at boot, but translation
+     * tables are NOT yet provided — strings remain in their authored Thai
+     * form. Future enhancement adds an i18n framework (react-i18next or
+     * similar) consuming this setting. OWNER editing today still affects
+     * accessibility readers via the lang attr.
+     */
+    language: 'th' | 'en';
+    /**
+     * D1.1.2.1 — DocumentType → 2-4 letter prefix mapping. Used by the UI to
+     * render document number badges (e.g. show "EX" next to an EXPENSE doc).
+     * Always returns the full default mapping when no override is configured.
+     */
+    docPrefixMap: Record<DocumentType, string>;
+    /**
+     * D1.3.6.2 — default-tick preference for the SettlementLinesSection bill
+     * list. Whitelist of three modes:
+     *   - `'all'`            — every fetched bill pre-ticked
+     *   - `'none'`           — nothing pre-ticked (manual selection only)
+     *   - `'overdue_only'`   — only bills past their `documentDate`-derived
+     *                          due date pre-ticked (default)
+     * Anything outside the whitelist (mis-edit, legacy value, malformed
+     * string) falls back to `'overdue_only'`.
+     */
+    settlementDefaultTick: 'all' | 'none' | 'overdue_only';
+    /**
+     * D1.1.3.2 — configurable WHT-rate dropdown. Always at least the 5
+     * defaults (1/3/5/10/15 %). Each entry may carry an optional
+     * `effectiveDate` (D1.1.3.5) — frontend filters out future-dated
+     * entries when rendering the picker.
+     */
+    whtRates: { rate: number; label: string; effectiveDate?: string | null }[];
+    /**
+     * D1.3.3.1 — global toggle for data-export endpoints (Excel/PDF/CSV).
+     * Default true. When OWNER sets `export_enabled = 'false'`, the server
+     * blocks PDF export endpoints with HTTP 403 and the web hides the
+     * "ส่งออก Excel" / "ดาวน์โหลด PDF" buttons. Useful for compliance
+     * lockdown periods (e.g. statutory audit window) where uncontrolled
+     * data extraction needs to be paused.
+     */
+    exportEnabled: boolean;
+    /**
+     * D1.4.3.2 — gate the weekly audit-log archive sweep
+     * (`AuditRetentionCron.archiveOldEntries`). Default `true`. When `false`,
+     * the cron skips without touching rows. Hard-delete remains impossible
+     * regardless (BEFORE DELETE trigger on audit_logs), so flipping this off
+     * just keeps rows in the hot set rather than purging the legal trail.
+     */
+    auditLogArchiveEnabled: boolean;
+    /**
+     * D1.4.3.4 — preferred format for data export / compliance backup.
+     * Whitelist `'JSON'` / `'CSV'` / `'XLSX'`, default `'JSON'`. Existing
+     * export buttons across the app should select this as the DEFAULT
+     * option in their format dropdown; the user can still override per
+     * export. Future automated compliance-backup jobs should consume
+     * this value via `getUiFlags()` rather than re-reading the
+     * SystemConfig row directly.
+     */
+    dataExportFormat: 'JSON' | 'CSV' | 'XLSX';
+    /**
+     * D1.4.3.6 — gate the LoginAuditLog row INSERT in
+     * `LoginAuditService.record`. Default `true`. When `false`, no audit
+     * row is written for login attempts — known-device tracking +
+     * new-device LINE alerts still run (security alerting independent of
+     * audit retention) and failed-attempt counting + account lockout in
+     * `AuthService` are unaffected (those drive the v3 account-lockout
+     * hardening, NOT the audit trail).
+     */
+    loginLogEnabled: boolean;
+    /**
+     * D1.3.4.2 — days threshold for the SAMEDAY→ACCRUAL auto-switch.
+     * Default `0` = any past document date triggers the flip (preserves
+     * the pre-Phase-4 hardcoded behavior). When set to N>0, the flip only
+     * fires when `(today − documentDate) > N` days. Useful when the shop
+     * routinely books cash purchases the next day (set `1` to tolerate a
+     * one-day lag without flipping to ACCRUAL). Clamped to 0–30 on read;
+     * non-integer / NaN / negative values fall back to 0.
+     *
+     * Originally marked SKIP per Phase 2 decision report; shipped per
+     * owner directive 2026-05-17 to reach 100% A1 coverage.
+     */
+    smartSwitchThresholdDays: number;
+    /**
+     * D1.3.4.1 — gate the auto SAMEDAY→ACCRUAL switch logic in the expense
+     * entry form. Default `true` preserves the existing one-way auto-flip
+     * (ExpenseFormV4: when the user picks a past `documentDate` while
+     * docType is SAMEDAY, flip to ACCRUAL). When `false` the auto-flip is
+     * skipped and the user must manually pick SAMEDAY vs ACCRUAL — useful
+     * for accountants who explicitly want SAMEDAY entries with a backdated
+     * invoice (e.g. cash purchases booked next day).
+     *
+     * Originally marked SKIP per Phase 2 decision report; shipped per
+     * owner directive 2026-05-17 to reach 100% A1 coverage.
+     */
+    smartDoctypeSwitchEnabled: boolean;
+    /**
+     * D1.1.6.3 — auto-route the ≤1฿ rounding remainder on Payment receipts
+     * to adj_underpay (52-1104) / adj_overpay (53-1503). Default TRUE.
+     * When FALSE, PaymentReceipt2B + PaymentReceipt2B-split throw
+     * BadRequestException on any non-zero rounding diff, forcing a manual
+     * JV to clear the residual. Exposed here so the admin Settings UI can
+     * render the toggle; the actual server-side enforcement lives in the JE
+     * templates (they read `adj_auto_route` directly via PrismaService).
+     */
+    adjAutoRoute: boolean;
+    /**
+     * D1.3.6.1 — max number of bills (cleared docs) per VENDOR_SETTLEMENT
+     * document. Default 100 (matches the legacy `limit=100` literal that
+     * `SettlementLinesSection.tsx` used to pull from `/expense-documents`).
+     * Clamped to 1–500 on read so an OWNER mis-edit can't disable the cap or
+     * blow up the SE form. Server enforces the cap on `createSettlement()`;
+     * UI uses the value to surface an early-warning banner before submit.
+     */
+    settlementMaxBillsPerDoc: number;
+    /**
+     * D1.1.5.4 — Petty Cash replenish alert threshold (THB). Default 5000,
+     * valid 0–50000 (clamp). When the running float balance falls below this
+     * number, `PettyCashReplenishAlertCron` (daily 09:00 BKK) notifies all
+     * active OWNERs via IN_APP. Setting to 0 disables the alert entirely
+     * (kill switch — owner can pick "dead" semantics by flipping this to 0).
+     */
+    pettyCashReplenishThreshold: number;
+    /**
+     * D1.1.5.1 — Petty Cash feature flag. Default true (feature is shipped
+     * and active per PRs #867+#868). When OWNER sets `petty_cash_enabled=false`,
+     * the web UI hides the Petty Cash doc-type card + the section in
+     * ExpenseFormV4, and the backend `createPettyCash` rejects with BadRequest
+     * "ระบบเงินสดย่อยถูกปิดใช้งาน".
+     */
+    pettyCashEnabled: boolean;
+    /**
+     * D1.2.5.3 — render the 3-column partial-payment breakdown (ยอดเดิม /
+     * ยอดที่ชำระ / ยอดคงเหลือ) on the voucher. Default true. When false the
+     * voucher shows only a single "ยอดที่ชำระ" column.
+     */
+    voucherShowPartialColumns: boolean;
+    /**
+     * D1.2.5.2 — include the rounding-adjustment / overpay-adjustment journal
+     * lines (52-1104, 53-1503) in the **printable** voucher layout. Default
+     * true. When false the rows are still rendered on screen so the JE
+     * preview stays complete, but the print stylesheet hides them so the
+     * physical paper voucher doesn't show the adjustment cents.
+     */
+    voucherIncludeAdjustment: boolean;
+    /**
+     * D1.2.5.1 — voucher print mode.
+     *   - 'multi' (default) — emits BOTH the original (ต้นฉบับ) and the
+     *     customer-copy (สำเนา) sheets, each on its own A4 page. The
+     *     customer copy carries a "สำเนา" watermark in the header.
+     *   - 'single' — renders only the original sheet.
+     * Whitelisted; unknown values fall back to 'multi'.
+     */
+    voucherPrintMode: 'single' | 'multi';
+    /**
+     * D1.2.4.1 — global toggle for the Expense Templates feature. When false,
+     * ExpenseTemplatesService rejects all writes (create/update/delete/
+     * instantiate) with a 403 ForbiddenException, and the UI hides the
+     * "บันทึกเป็นรายการโปรด" buttons + templates list. List/read endpoints
+     * still resolve (so legacy data isn't hidden) but new writes are blocked.
+     * Default true to preserve current behaviour.
+     */
+    templatesEnabled: boolean;
+    /**
+     * D1.2.4.2 — per-user quota of saved Expense Templates. Default 20.
+     * Clamped to 1–1000 on read. `ExpenseTemplatesService.create` counts
+     * the caller's existing (non-deleted) templates against this cap and
+     * rejects with BadRequestException ("โควต้าเทมเพลตเต็มแล้ว — ลบ
+     * เทมเพลตเก่าก่อนสร้างใหม่") when count >= cap. UI surfaces it as a
+     * "X/N" badge on the favorites picker for at-a-glance awareness.
+     */
+    maxTemplatesPerUser: number;
+    /**
+     * D1.2.4.4 — gates the `{{variable}}` interpolation feature on
+     * Expense Templates. Default true. When false, the UI hides the
+     * "ใส่ตัวแปร" affordance and `interpolateTemplate()` callers should
+     * skip interpolation. The util itself (`template-interpolation.util`)
+     * is pure and always available; this flag controls whether the
+     * product surfaces the feature to users.
+     */
+    templateVariablesEnabled: boolean;
+    /**
+     * D1.2.3.5 — thousands separator style for the generic number formatter.
+     * Whitelisted: 'comma' (1,234,567) default, 'space' (1 234 567), or
+     * 'none' (1234567). SystemConfig key `thousands_separator`. Invalid
+     * values fall back to 'comma'. Voucher-specific helpers in `lib/date.ts`
+     * are unaffected — they format dates not numbers.
+     */
+    thousandsSeparator: 'comma' | 'space' | 'none';
+    /**
+     * D1.2.3.4 — default number of decimal places for the generic number
+     * formatter. Integer 0-4 inclusive. Default 2 (Thai currency convention).
+     * SystemConfig key `decimal_places`. Out-of-range/non-integer values
+     * clamp to default. The pref is the *default only* — `formatNumberDecimal`
+     * call sites that explicitly pass a digit count continue to work as
+     * before, preserving backwards compat.
+     */
+    decimalPlaces: number;
+    /**
+     * D1.2.3.3 — date display format toggle: 'BE' = Buddhist Era (พ.ศ., +543)
+     * default, 'CE' = Christian/Common Era (Gregorian ค.ศ.). SystemConfig
+     * key `date_format`. Applies to the *generic* `formatDateShort` family
+     * in `apps/web/src/utils/formatters.ts`. Voucher-specific helpers
+     * (`formatThaiDate*` in `apps/web/src/lib/date.ts`) remain BE-only by
+     * design — legal/tax documents always show พ.ศ. regardless of UI pref.
+     */
+    dateFormat: 'BE' | 'CE';
+    /**
+     * D1.2.1.1 — Approval Workflow opt-in toggle. Default `false` (legacy
+     * lifecycle DRAFT → POSTED is preserved). When `true`, expense docs
+     * follow DRAFT → PENDING_APPROVAL → APPROVED → POSTED. The submit /
+     * approve actions live on `ExpenseDocumentsService.submitForApproval()`
+     * + `.approve()` (D1.2.1.6). Threshold + doc-type filters are layered
+     * on top by D1.2.1.2 / D1.2.1.4. Web UI uses this flag to conditionally
+     * render the "ส่งขออนุมัติ" button instead of "Post".
+     */
+    approvalEnabled: boolean;
+    /**
+     * D1.2.1.2 — threshold (THB) above which docs require approval. Surfaced
+     * to the UI for helper-text only. Backend is the source of truth.
+     */
+    approvalThreshold: number;
+    /**
+     * D1.2.1.3 — user IDs that may approve PENDING_APPROVAL docs (besides
+     * OWNER). Frontend gates "อนุมัติเอกสาร" button on this.
+     */
+    approversList: string[];
+    /**
+     * D1.2.1.4 — document types that always require approval (OR-composed
+     * with threshold). Default `['PAYROLL']`.
+     */
+    approvalRequiredDocTypes: string[];
+    /**
+     * D1.2.1.5 — fan out IN_APP notifications when a doc enters
+     * PENDING_APPROVAL. Default `true`. Respects the master gate
+     * `in_app_notifications_enabled` (D1.3.1.4) — if either is off, no
+     * notifications are sent. Read by `ExpenseDocumentsService.notifyApprovers`.
+     */
+    notificationOnPending: boolean;
+    /**
+     * D1.2.3.2 — default pagination size for list pages. Valid integer 10-200
+     * inclusive. Default `50`. Out-of-range or non-numeric values clamp to
+     * the default so a malformed admin edit can't break list pages. SystemConfig
+     * key `pagination_size`. Frontend uses this as the default `limit` query
+     * param; URL `?size=N` overrides per-session, but new sessions start at
+     * this value.
+     */
+    paginationSize: number;
+    /**
+     * D1.2.3.1 — default time range preset selected on list-page mount.
+     * Whitelisted: `'all' | 'this_month' | 'last_month'`. Default
+     * `'this_month'` (matches accounting workflow expectation). Page-level
+     * code uses this to initialize startDate/endDate when no URL query
+     * params override; user changes mid-session are NOT persisted to the
+     * setting — this is the *initial* state only.
+     */
+    defaultTimeRange: 'all' | 'this_month' | 'last_month';
+    /**
+     * D1.3.5.1 — default time range preset for the expense daily-summary
+     * page. Whitelisted: `'today' | 'this_week' | 'this_month' | 'last_month'`.
+     * Default `'this_month'`. Wider preset set than `defaultTimeRange` because
+     * the summary page is consumed daily (operations) AND monthly (ผู้จัดการ
+     * รีวิว). Page-level code uses this to initialize startDate/endDate on
+     * mount; mid-session user changes are NOT persisted to the setting.
+     */
+    summaryDefaultRange: 'today' | 'this_week' | 'this_month' | 'last_month';
+    /**
+     * D1.3.1.2 — AP-due alerts cron toggle. Default `false` (OFF) until
+     * ExpenseDocument has a real `dueDate` column — currently the cron uses
+     * `documentDate` as a proxy, so leaving it ON by default would spam
+     * approvers with daily alerts for every POSTED doc 3+ days old regardless
+     * of actual vendor credit terms. OWNERs opt-in via /settings when they
+     * accept the documentDate semantics.
+     */
+    apDueAlertsEnabled: boolean;
+    /**
+     * D1.3.1.2 — number of days past `documentDate` that a POSTED unpaid
+     * expense doc must reach before triggering an AP-due alert. Default 3.
+     * NOTE: ExpenseDocument has no explicit dueDate column today; the cron
+     * approximates "due in N days" by counting days since `documentDate`.
+     */
+    apDueDaysBefore: number;
+    /**
+     * D1.3.1.1 — DRAFT alerts toggle. When true, `DraftAlertsCron` (daily at
+     * 09:00 BKK) scans expense docs in DRAFT for longer than the configured
+     * threshold and sends an in-app notification to the creator. Default
+     * `false` (opt-in) so existing deploys don't suddenly start spamming
+     * users when this PR rolls out.
+     */
+    draftAlertsEnabled: boolean;
+    /**
+     * D1.3.1.1 — DRAFT alert threshold in days. Drafts older than this trigger
+     * the alert. Default 7. Validated `> 0`.
+     */
+    draftAlertThresholdDays: number;
+    /**
+     * D1.1.6 — adjustment account codes for the V4 multi-line Adjustment
+     * row. Frontend `AdjustmentSection.tsx` previously hardcoded
+     * '52-1104' / '53-1503'; now reads from this flag so OWNER can rebind
+     * the codes (e.g. branch-specific CoA variant) without a frontend
+     * deploy. Defaults preserve the legacy behaviour.
+     *
+     * `underpay` is suggested when amountPaid < expected (Dr side).
+     * `overpay` is suggested when amountPaid > expected (Cr side).
+     *
+     * Backend JE templates (PaymentReceipt2BTemplate, etc.) still
+     * hardcode these codes — keeping this scoped to the V4 form so
+     * server-side templates stay deterministic for golden CSV.
+     */
+    adjustmentCodes: { underpay: string; overpay: string };
+    /**
+     * D1.4.1.1 — BOOTSTRAP default for sidebar collapse on a brand-new device
+     * (no `sidebar_collapse` key in localStorage). Once the user toggles the
+     * sidebar in the UI, their personal preference is persisted and takes
+     * precedence — this flag never overrides an existing per-user value.
+     * Default false (= expanded). OWNER stores 'true' / 'false'.
+     */
+    sidebarCollapsedDefault: boolean;
+    /**
+     * D1.4.1.2 — controls whether keyboard-shortcut hints (the Shift+? help
+     * dialog binding + per-item kbd hints) are exposed to the user. Default
+     * true preserves the existing UX. OWNER stores 'true'/'false'.
+     */
+    showKeyboardShortcuts: boolean;
+    /**
+     * D1.4.1.3 — global animations + transitions toggle. Default true.
+     * When false, the web app sets `data-animations-disabled="true"` on
+     * `<html>` and a CSS rule strips `transition` / `animation` properties.
+     * Complements the OS-level `prefers-reduced-motion` media query for
+     * accessibility users on browsers that don't expose the OS setting.
+     */
+    animationEnabled: boolean;
+    /**
+     * D1.3.1.4 — Master IN_APP notification kill switch. Default `true`.
+     * When `false`, NotificationsService.send() returns
+     * `{ id: '', status: 'SKIPPED', blockReason: 'IN_APP_DISABLED' }` for
+     * `IN_APP` channel calls — no DB write, no exception. LINE/SMS unaffected.
+     * Surfaced here so UIs can render banners explaining the silent skip.
+     */
+    inAppNotificationsEnabled: boolean;
+    /**
+     * D1.4.1.4 — BOOTSTRAP default theme for first-time devices (no `theme`
+     * key in localStorage from next-themes). 'system' default = respect OS
+     * `prefers-color-scheme`. Existing per-user preference always wins after
+     * the user has clicked the theme toggle once.
+     */
+    darkModeDefault: 'light' | 'dark' | 'system';
+    /**
+     * D1.4.2.1 — long-running query timeout (seconds). Default 30, valid
+     * range 5–300 (clamped). Currently INFORMATIONAL — applying this to
+     * Postgres requires a DB-level `statement_timeout` setting, and the
+     * shared axios client in `apps/web/src/lib/api.ts` uses a fixed 15s
+     * timeout. The flag is exposed so OWNER can advertise the intended
+     * cutoff to operators and a future PR can wire it into either the
+     * axios `timeout` field or a Postgres `SET LOCAL statement_timeout`
+     * pre-hook. Originally SKIP per Phase 2; shipped per owner directive
+     * 2026-05-17 to reach 100% A1 coverage.
+     */
+    queryTimeoutSeconds: number;
+    /**
+     * D1.3.1.3 — active email provider. Whitelisted: 'smtp' (default —
+     * uses SMTP_HOST/PORT/USER/PASS env vars) / 'sendgrid' (stub — owner
+     * must wire SENDGRID_API_KEY before flipping). UI shows "Current
+     * provider: SMTP / Sendgrid" + warns when SMTP env not configured.
+     */
+    emailProvider: 'smtp' | 'sendgrid';
+    /**
+     * D1.4.2.2 — react-query staleTime (seconds) for dashboard KPI / chart
+     * queries. Default 60s, valid 10–3600 (clamped). Actively wired into
+     * `DashboardPage`'s `dashboardStaleTime` so OWNER can balance
+     * freshness vs DB cost without redeploy. Originally SKIP per Phase 2;
+     * shipped per owner directive 2026-05-17 to reach 100% A1 coverage.
+     */
+    cacheTtlDashboard: number;
+    /**
+     * D1.4.2.3 — react-query staleTime (seconds) for aggregated report
+     * queries (P&L, monthly P&L, trial balance). Default 300s, valid
+     * 30–7200 (clamped). Wired into `ProfitLossPage`.
+     */
+    cacheTtlReports: number;
+    /**
+     * D1.3.6.3 — allow per-line partial settlement on VENDOR_SETTLEMENT.
+     * Default true (V12 adjustment logic remains active). When false the
+     * server rejects any line where `amountSettled < remainingCap` (±0.01
+     * rounding slop) — every line must clear the full outstanding balance.
+     * The web app additionally disables the "จำนวนที่จ่าย" input column so
+     * the user can't paste in a partial figure.
+     */
+    settlementPartialPaymentEnabled: boolean;
+    /**
+     * D1.3.2.1 — VIEWER role activation flag. Default false (Q4-gated).
+     * When true, future guards/widening code can extend @Roles() lists on
+     * expense / other-income / asset modules to include the VIEWER role.
+     * Schema enum value always exists (UserRole.VIEWER) so the flip is
+     * SystemConfig-only; no migration needed to roll forward/back.
+     */
+    viewerRoleEnabled: boolean;
+    /**
+     * D1.3.3.3 — outbound webhook dispatch master switch. **DEFAULT-OFF** per
+     * accountant package. When false, `WebhooksService.dispatchEvent` is a
+     * no-op and the OWNER's `POST /webhooks/test/:id` returns 400. Inbound
+     * webhooks (paysolutions / sms / line / facebook) are NOT gated by this
+     * flag — they run on dedicated controllers and remain critical for
+     * payment processing.
+     */
+    webhooksEnabled: boolean;
+    /**
+     * D1.3.2.2 — dynamic bundle name controlling who can access Settings.
+     * Whitelisted: `'OWNER'` (default) / `'OWNER+FINANCE_MANAGER'` /
+     * `'OWNER+ACCOUNTANT'` / `'OWNER+ALL'`. Read at request time by
+     * `SettingsAccessGuard`; exposed here so the web app can render an
+     * informational badge on /settings. Invalid SystemConfig values fall
+     * back to `'OWNER'` (preserves OWNER-only behavior).
+     */
+    settingsAccessRole: 'OWNER' | 'OWNER+FINANCE_MANAGER' | 'OWNER+ACCOUNTANT' | 'OWNER+ALL';
+    /**
+     * D1.3.2.3 — dynamic bundle controlling who can POST expense documents
+     * (DRAFT → ACCRUAL). Whitelisted: `'OWNER+FINANCE_MANAGER+ACCOUNTANT'`
+     * (default — current behavior) / `'OWNER+FINANCE_MANAGER'` /
+     * `'OWNER_ONLY'` / `'OWNER+ALL_NON_SALES'`. Enforced at request time by
+     * `PostPermissionGuard`; exposed here so UIs can hide the "Post" button
+     * for roles that won't be allowed (defence against confusing 403s).
+     */
+    postPermission:
+      | 'OWNER+FINANCE_MANAGER+ACCOUNTANT'
+      | 'OWNER+FINANCE_MANAGER'
+      | 'OWNER_ONLY'
+      | 'OWNER+ALL_NON_SALES';
+    /**
+     * D1.3.2.4 + InternalControlActionBar — dynamic bundle controlling who
+     * can reverse/void accounting documents. Whitelisted:
+     *   - `'OWNER_ONLY'`
+     *   - `'OWNER+FINANCE_MANAGER'` (default — legacy behavior)
+     *   - `'OWNER+FINANCE_MANAGER+ACCOUNTANT'`
+     *   - `'CUSTOM'` — per-user via `User.canReverseOverride`
+     * Enforced at request time by `ReversePermissionGuard`; exposed here so
+     * UIs can hide the "Void/Reverse" button for roles that won't be allowed.
+     */
+    reversePermission:
+      | 'OWNER_ONLY'
+      | 'OWNER+FINANCE_MANAGER'
+      | 'OWNER+FINANCE_MANAGER+ACCOUNTANT'
+      | 'CUSTOM';
+  }> {
+    const taxExemptWarningEnabled = await this.readBoolean(
+      'TAX_EXEMPT_WARNING_ENABLED',
+      true,
+    );
+    const reverseReasonRequired = await this.readBoolean(
+      'reverse_reason_required',
+      true,
+    );
+    const reverseReasons = await this.getReverseReasons();
+    const paymentDateAllowFuture = await this.readBoolean(
+      'payment_date_allow_future',
+      true,
+    );
+    const voucherShowQrCode = await this.readBoolean('voucher_show_qr_code', true);
+    // D1.2.2.6 — language. Whitelist 'th' / 'en'; everything else → 'th'.
+    const language = pickEnum(await this.getKey('language'), ['th', 'en'] as const, 'th');
+    // D1.1.2.1 — DocumentType → prefix map (defaults applied per type when
+    // stored value is missing or malformed).
+    const docPrefixMap = await this.getDocPrefixMap();
+    // D1.3.6.2 — settlement_default_tick. Whitelist 'all' / 'none' /
+    // 'overdue_only'; everything else → 'overdue_only' (spec default).
+    const settlementDefaultTick = pickEnum(
+      await this.getKey('settlement_default_tick'),
+      ['all', 'none', 'overdue_only'] as const,
+      'overdue_only',
+    );
+    // D1.1.3.2 — WHT rates (default 5 canonical entries + optional D1.1.3.5
+    // effectiveDate per entry).
+    const whtRates = await this.getWhtRates();
+    // D1.3.3.1 — export_enabled. Default true.
+    const exportEnabled = await this.readBoolean('export_enabled', true);
+    // D1.4.3.2 — audit log archive toggle. Default true.
+    const auditLogArchiveEnabled = await this.readBoolean(
+      'audit_log_archive_enabled',
+      true,
+    );
+    // D1.4.3.4 — data_export_format. Whitelist JSON/CSV/XLSX; default JSON.
+    const dataExportFormat = pickEnum(
+      await this.getKey('data_export_format'),
+      ['JSON', 'CSV', 'XLSX'] as const,
+      'JSON',
+    );
+    // D1.4.3.6 — login_log_enabled. Default true.
+    const loginLogEnabled = await this.readBoolean('login_log_enabled', true);
+    // D1.3.4.2 — smart-switch threshold (days). Clamp 0–30; non-integer /
+    // NaN / negative → 0. Default 0 = legacy behavior (any past date flips).
+    const smartSwitchThresholdDays = clampInt(
+      await this.readNumber('smart_switch_threshold_days', 0),
+      0,
+      30,
+      0,
+    );
+    // D1.3.4.1 — smart_doctype_switch_enabled (default true).
+    const smartDoctypeSwitchEnabled = await this.readBoolean(
+      'smart_doctype_switch_enabled',
+      true,
+    );
+    // D1.1.6.3 — adj_auto_route. Defaults TRUE so first-boot behaviour
+    // mirrors the original auto-route-to-52-1104/53-1503 logic.
+    const adjAutoRoute = await this.readBoolean('adj_auto_route', true);
+    // D1.3.6.1 — settlement_max_bills_per_doc. Clamp to 1–500 inclusive;
+    // anything outside (incl. NaN / negative) falls back to the default 100
+    // which matches the previous hardcoded limit.
+    const settlementMaxBillsPerDoc = clampInt(
+      await this.readNumber('settlement_max_bills_per_doc', 100),
+      1,
+      500,
+      100,
+    );
+    // D1.1.5.4 — Petty Cash replenish threshold. Default 5000, valid 0–50000.
+    // Negative or NaN silently clamps to default 5000 so a bad SystemConfig
+    // row can't accidentally suppress the alert via negative comparison.
+    const thresholdRaw = await this.readNumber('petty_cash_replenish_threshold', 5000);
+    let pettyCashReplenishThreshold = thresholdRaw;
+    if (!Number.isFinite(pettyCashReplenishThreshold) || pettyCashReplenishThreshold < 0) {
+      pettyCashReplenishThreshold = 5000;
+    } else if (pettyCashReplenishThreshold > 50000) {
+      pettyCashReplenishThreshold = 50000;
+    }
+    // D1.1.5.1 — Petty Cash feature flag. Default true (feature shipped).
+    const pettyCashEnabled = await this.readBoolean('petty_cash_enabled', true);
+    // D1.2.5.3 — show ยอดเดิม / ยอดที่ชำระ / ยอดคงเหลือ on voucher.
+    const voucherShowPartialColumns = await this.readBoolean(
+      'voucher_show_partial_columns',
+      true,
+    );
+    // D1.2.5.2 — include adjustment rows (52-1104 / 53-1503) on printed voucher.
+    const voucherIncludeAdjustment = await this.readBoolean(
+      'voucher_include_adjustment',
+      true,
+    );
+    // D1.2.5.1 — voucher print mode. Whitelist 'single' / 'multi'; default 'multi'.
+    const voucherPrintMode = pickEnum(
+      await this.getKey('voucher_print_mode_default'),
+      ['single', 'multi'] as const,
+      'multi',
+    );
+    // D1.2.4.1 — Expense Templates feature flag.
+    const templatesEnabled = await this.readBoolean('templates_enabled', true);
+    // D1.2.4.2 — per-user template quota. Clamp to 1–1000 (same range as
+    // ExpenseTemplatesService's internal cap-reader so values stay aligned).
+    // D1.2.4.4 — gate the {{variable}} interpolation surface. Default true.
+    const templateVariablesEnabled = await this.readBoolean(
+      'template_variables_enabled',
+      true,
+    );
+    // D1.2.4.2 — per-user template quota. Default 20, clamp to 1–1000 on
+    // read so a bad SystemConfig row can't disable the cap (which would
+    // let one user balloon the favorites table) or pin it to 0 (which
+    // would lock everyone out of saving new templates).
+    const maxTemplatesPerUserRaw = await this.readNumber('max_templates_per_user', 20);
+    const maxTemplatesPerUser =
+      Number.isFinite(maxTemplatesPerUserRaw) && maxTemplatesPerUserRaw >= 1
+        ? Math.min(Math.floor(maxTemplatesPerUserRaw), 1000)
+        : 20;
+    // D1.2.3.5 — thousands_separator. Whitelist 'comma' / 'space' / 'none';
+    // everything else → 'comma'.
+    const thousandsSeparator = pickEnum(
+      await this.getKey('thousands_separator'),
+      ['comma', 'space', 'none'] as const,
+      'comma',
+    );
+    // D1.2.3.4 — decimal_places. Integer 0-4 inclusive; clamp to 2 default
+    // for out-of-range / non-integer values.
+    const decimalPlaces = clampInt(await this.readNumber('decimal_places', 2), 0, 4, 2);
+    // D1.2.3.3 — date_format. Whitelist 'BE' / 'CE'; everything else → 'BE'.
+    // Default 'BE' so existing flows are unchanged.
+    const dateFormat = pickEnum(await this.getKey('date_format'), ['BE', 'CE'] as const, 'BE');
+    // D1.2.1.1 — Approval Workflow opt-in. Default true per
+    // Settings_Audit_Core_v2.0.md spec. Owner can flip to `false` via
+    // SystemConfig if rollout needs to be gradual.
+    const approvalEnabled = await this.readBoolean('approval_enabled', true);
+    // D1.2.1.2 — approval threshold (THB). Default 0; clamp negatives.
+    const approvalThresholdRaw = await this.readNumber('approval_threshold', 0);
+    const approvalThreshold = approvalThresholdRaw >= 0 ? approvalThresholdRaw : 0;
+    // D1.2.1.3 — approvers_list (JSON array of user IDs). Empty default.
+    const approversList = parseStringArray(await this.getKey('approvers_list'), []);
+    // D1.2.1.4 — approval_required_doc_types (JSON array). Default ['PAYROLL'].
+    const approvalRequiredDocTypes = parseStringArray(
+      await this.getKey('approval_required_doc_types'),
+      ['PAYROLL'],
+      { requireNonEmpty: true },
+    );
+    // D1.2.1.5 — notification fan-out toggle. Default true. Respects master
+    // gate `in_app_notifications_enabled` downstream in NotificationsService.
+    const notificationOnPending = await this.readBoolean('notification_on_pending', true);
+    // D1.2.3.2 — pagination_size. Integer 10-200 inclusive; clamp to 50
+    // default for out-of-range or non-numeric values so list pages remain
+    // usable even when SystemConfig is mis-edited.
+    const paginationSize = clampInt(await this.readNumber('pagination_size', 50), 10, 200, 50);
+    // D1.2.3.1 — default time range. Whitelist; everything else falls
+    // through to 'this_month' so a malformed admin edit can't break list
+    // pages that depend on this for initial state.
+    const defaultTimeRange = pickEnum(
+      await this.getKey('default_time_range'),
+      ['all', 'this_month', 'last_month'] as const,
+      'this_month',
+    );
+    // D1.3.5.1 — summary-page default range. Wider whitelist than
+    // defaultTimeRange (no 'all' — summary always wants a bounded period,
+    // and the page UI doesn't currently expose an "all" option; the
+    // companion D1.3.5.2 banner explains gracefully if 'all' is ever wired).
+    const summaryDefaultRange = pickEnum(
+      await this.getKey('summary_default_range'),
+      ['today', 'this_week', 'this_month', 'last_month'] as const,
+      'this_month',
+    );
+    // D1.3.1.2 — AP-due alerts. Default OFF until ExpenseDocument has a real
+    // dueDate column (documentDate proxy would otherwise spam every POSTED doc).
+    const apDueAlertsEnabled = await this.readBoolean('ap_due_alerts_enabled', false);
+    const apDueDaysBeforeRaw = await this.readNumber('ap_due_days_before', 3);
+    const apDueDaysBefore = apDueDaysBeforeRaw >= 0 ? apDueDaysBeforeRaw : 3;
+    // D1.3.1.1 — DRAFT alerts (opt-in, default off).
+    const draftAlertsEnabled = await this.readBoolean('draft_alerts_enabled', false);
+    const draftAlertThresholdDaysRaw = await this.readNumber(
+      'draft_alert_threshold_days',
+      7,
+    );
+    const draftAlertThresholdDays =
+      draftAlertThresholdDaysRaw > 0 ? draftAlertThresholdDaysRaw : 7;
+    // D1.1.6 — adjustment codes for the V4 form's manual reconciliation
+    // row. Codes must match the format `\d{2}-\d{4}` to be accepted;
+    // anything malformed falls back to the legacy default. Empty string
+    // never returned — defensive against half-saved SystemConfig rows.
+    const isValidCode = (raw: string | null): raw is string =>
+      !!raw && /^\d{2}-\d{4}$/.test(raw);
+    const underpayRaw = await this.getKey('adjustment_code_underpay');
+    const overpayRaw = await this.getKey('adjustment_code_overpay');
+    const adjustmentCodes = {
+      underpay: isValidCode(underpayRaw) ? underpayRaw : '52-1104',
+      overpay: isValidCode(overpayRaw) ? overpayRaw : '53-1503',
+    };
+    // D1.4.1.1 — sidebar bootstrap default. `readBoolean` already whitelists
+    // 'true' / 'false' / '1' / '0' so a bad row falls back to false (expanded).
+    const sidebarCollapsedDefault = await this.readBoolean(
+      'sidebar_collapsed_default',
+      false,
+    );
+    // D1.4.1.2 — keyboard shortcut hints + help-dialog binding. Default true.
+    const showKeyboardShortcuts = await this.readBoolean(
+      'show_keyboard_shortcuts',
+      true,
+    );
+    // D1.4.1.3 — animation enabled. Default true.
+    const animationEnabled = await this.readBoolean('animation_enabled', true);
+    // D1.3.1.4 — Master IN_APP notification toggle. Default true.
+    const inAppNotificationsEnabled = await this.readBoolean(
+      'in_app_notifications_enabled',
+      true,
+    );
+    // D1.4.1.4 — dark_mode_default. Whitelist light/dark/system, default system.
+    const darkModeDefault = pickEnum(
+      await this.getKey('dark_mode_default'),
+      ['light', 'dark', 'system'] as const,
+      'system',
+    );
+    // D1.4.2.1 — query_timeout_seconds. Clamp to [5, 300].
+    const queryTimeoutSeconds = clampFloat(
+      await this.readNumber('query_timeout_seconds', 30),
+      5,
+      300,
+      30,
+    );
+    // D1.3.1.3 — email provider whitelist. Default smtp.
+    const emailProvider = pickEnum(
+      await this.getKey('email_provider'),
+      ['smtp', 'sendgrid'] as const,
+      'smtp',
+    );
+    // D1.4.2.2 — cache_ttl_dashboard. Clamp to [10, 3600] seconds.
+    const cacheTtlDashboard = clampFloat(
+      await this.readNumber('cache_ttl_dashboard', 60),
+      10,
+      3600,
+      60,
+    );
+    // D1.4.2.3 — cache_ttl_reports. Clamp to [30, 7200] seconds.
+    const cacheTtlReports = clampFloat(
+      await this.readNumber('cache_ttl_reports', 300),
+      30,
+      7200,
+      300,
+    );
+    // D1.3.6.3 — settlement_partial_payment_enabled. Default true (V12
+    // adjustments remain active). When false, web app disables the partial
+    // amount input + server rejects underpaid lines.
+    const settlementPartialPaymentEnabled = await this.readBoolean(
+      'settlement_partial_payment_enabled',
+      true,
+    );
+    // D1.3.2.1 — VIEWER role activation. Conservative default false.
+    const viewerRoleEnabled = await this.readBoolean('viewer_role_enabled', false);
+    // D1.3.3.3 — webhooks_enabled. DEFAULT-OFF per accountant package.
+    const webhooksEnabled = await this.readBoolean('webhooks_enabled', false);
+    // D1.3.2.2 — settings_access_role. Whitelist 4 values; everything else
+    // (missing key, malformed value) falls back to 'OWNER' so behavior
+    // matches the pre-Phase-2 hardcoded @Roles('OWNER').
+    const settingsAccessRole = pickEnum(
+      await this.getKey('settings_access_role'),
+      ['OWNER', 'OWNER+FINANCE_MANAGER', 'OWNER+ACCOUNTANT', 'OWNER+ALL'] as const,
+      'OWNER',
+    );
+    // D1.3.2.3 — post_permission. Whitelist 4 values; everything else falls
+    // back to the default OWNER+FM+ACCOUNTANT bundle (matches the
+    // pre-Phase-2 static @Roles decorator).
+    const postPermission = pickEnum(
+      await this.getKey('post_permission'),
+      [
+        'OWNER+FINANCE_MANAGER+ACCOUNTANT',
+        'OWNER+FINANCE_MANAGER',
+        'OWNER_ONLY',
+        'OWNER+ALL_NON_SALES',
+      ] as const,
+      'OWNER+FINANCE_MANAGER+ACCOUNTANT',
+    );
+    // D1.3.2.4 + InternalControlActionBar — reverse_permission. Whitelist
+    // 4 values; everything else falls back to the default OWNER+FM bundle.
+    //   - OWNER_ONLY
+    //   - OWNER+FINANCE_MANAGER (default)
+    //   - OWNER+FINANCE_MANAGER+ACCOUNTANT
+    //   - CUSTOM (per-user via User.canReverseOverride, see ReversePermissionGuard)
+    const reversePermission = pickEnum(
+      await this.getKey('reverse_permission'),
+      ['OWNER_ONLY', 'OWNER+FINANCE_MANAGER', 'OWNER+FINANCE_MANAGER+ACCOUNTANT', 'CUSTOM'] as const,
+      'OWNER+FINANCE_MANAGER',
+    );
+    return {
+      taxExemptWarningEnabled,
+      reverseReasonRequired,
+      reverseReasons,
+      paymentDateAllowFuture,
+      voucherShowQrCode,
+      language,
+      docPrefixMap,
+      settlementDefaultTick,
+      whtRates,
+      exportEnabled,
+      auditLogArchiveEnabled,
+      dataExportFormat,
+      loginLogEnabled,
+      smartSwitchThresholdDays,
+      summaryDefaultRange,
+      smartDoctypeSwitchEnabled,
+      adjAutoRoute,
+      settlementMaxBillsPerDoc,
+      pettyCashReplenishThreshold,
+      pettyCashEnabled,
+      voucherShowPartialColumns,
+      voucherIncludeAdjustment,
+      voucherPrintMode,
+      templatesEnabled,
+      maxTemplatesPerUser,
+      templateVariablesEnabled,
+      thousandsSeparator,
+      decimalPlaces,
+      dateFormat,
+      approvalEnabled,
+      approvalThreshold,
+      approversList,
+      approvalRequiredDocTypes,
+      notificationOnPending,
+      paginationSize,
+      defaultTimeRange,
+      apDueAlertsEnabled,
+      apDueDaysBefore,
+      draftAlertsEnabled,
+      draftAlertThresholdDays,
+      adjustmentCodes,
+      sidebarCollapsedDefault,
+      showKeyboardShortcuts,
+      animationEnabled,
+      inAppNotificationsEnabled,
+      darkModeDefault,
+      queryTimeoutSeconds,
+      emailProvider,
+      cacheTtlDashboard,
+      cacheTtlReports,
+      settlementPartialPaymentEnabled,
+      viewerRoleEnabled,
+      webhooksEnabled,
+      settingsAccessRole,
+      postPermission,
+      reversePermission,
+    };
+  }
+
+  /**
+   * Typed bundle of collections-session tuning knobs. Defaults match the
+   * pre-Phase-2 hardcoded constants so first-boot behaviour is unchanged
+   * even if the SystemConfig rows haven't been seeded yet.
+   */
+  async getCollectionsConfig(): Promise<{
+    dailyCap: number;
+    workloadFloor: number;
+    etaPerContractMin: number;
+    sessionTargetMin: number;
+    selfClaimLockHours: number;
+  }> {
+    const dailyCap = await this.readNumber('collections.dailyCap', 30);
+    const workloadFloor = await this.readNumber('collections.workloadFloor', 10);
+    const etaPerContractMin = await this.readNumber('collections.etaPerContractMin', 5);
+    const sessionTargetMin = await this.readNumber('collections.sessionTargetMin', 150);
+    const selfClaimLockHours = await this.readNumber('collections.selfClaimLockHours', 2);
+    return { dailyCap, workloadFloor, etaPerContractMin, sessionTargetMin, selfClaimLockHours };
+  }
+}
