+++ apps/api/src/modules/settings/settings.service.spec.ts
@@ -1,5 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { SettingsService } from './settings.service';
+import { SettingsFlagsService } from './services/settings-flags.service';
+import { SettingsWriteService } from './services/settings-write.service';
+import { PettyCashCustodianService } from './services/petty-cash-custodian.service';
+import { DocNumberPreviewService } from './services/doc-number-preview.service';
 import { PrismaService } from '../../prisma/prisma.service';
 import { AuditService } from '../audit/audit.service';
 
@@ -24,6 +28,10 @@ describe('SettingsService audit trail', () => {
     const mod: TestingModule = await Test.createTestingModule({
       providers: [
         SettingsService,
+        SettingsFlagsService,
+        SettingsWriteService,
+        PettyCashCustodianService,
+        DocNumberPreviewService,
         { provide: PrismaService, useValue: prisma },
         { provide: AuditService, useValue: audit },
       ],
